backend_project/requirements.txt
flask>=3.0
PyJWT>=2.8
werkzeug>=3.0
gunicorn>=21.0
